# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/status/{job_id}":{"get":{"summary":"Get Job Status","description":"Returns the status of the given job_id\n\nArgs:\n    job_id: str: ID of the job whose status to check for\n\nReturns:\n    The status of the given job_id [\"completed\" OR \"error\" OR \"pending\"]","operationId":"get_job_status_status__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the job whose status to check for","title":"Job Id"},"description":"ID of the job whose status to check for"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/submit/{job_id}":{"post":{"summary":"Submit Job","description":"Returns the job_id after submitting it successfully\n\nArgs:\n    job_id: str: ID of the job to submit\n    delay_seconds: int: Proxy for indicating the time (in seconds) it\n    takes for the given job_id to complete successfully. (default 20)\n\nReturns: A string indicating successful submission of the job, with its ID","operationId":"submit_job_submit__job_id__post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the job to create","title":"Job Id"},"description":"ID of the job to create"},{"name":"delay_seconds","in":"query","required":false,"schema":{"type":"integer","description":"Time in seconds for the job to complete","default":20,"title":"Delay Seconds"},"description":"Time in seconds for the job to complete"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}